Added NPC that adjusts audio levels, "pause" only disables player input
Because I thought that it would be nice to have a "bard" in the Adventurer's Guild that you can talk to that also adjusts your audio levels. This involved changing some code related to cutscenes and dialogue.

It also involved making it easier to disable player input to the player actor, so I took this opportunity to also make the pause menu ONLY disable player input but not actually pause the game. That's right, you can't pause the game now! I thought it would be cooler this way.